+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [1.6.0] - 2025-05-05
+
 ### Added
 
 + Added an option to override Element's default username colors
@@ -77,7 +79,8 @@ and this project adheres to
 
 + Initial release
 
-[Unreleased]: https://github.com/mserajnik/element-emotes/compare/1.5.0...develop
+[Unreleased]: https://github.com/mserajnik/element-emotes/compare/1.6.0...develop
+[1.6.0]: https://github.com/mserajnik/element-emotes/compare/1.5.0...1.6.0
 [1.5.0]: https://github.com/mserajnik/element-emotes/compare/1.4.0...1.5.0
 [1.4.0]: https://github.com/mserajnik/element-emotes/compare/1.3.1...1.4.0
 [1.3.1]: https://github.com/mserajnik/element-emotes/compare/1.3.0...1.3.1